feat: confirm mail

+++ src/contexts/JWTContext.tsx
@@ -1,5 +1,6 @@
 'use client';
 
+import jwt from 'jsonwebtoken';
 import { createContext, PropsWithChildren, useEffect, useReducer, useState } from 'react';
 
 import accountReducer from 'store/accountReducer';
@@ -18,21 +19,42 @@ export const JWTProvider = ({ children }: PropsWithChildren) => {
   const [loading, setLoading] = useState(false);
   const [user, dispatch] = useReducer(accountReducer, null);
   const [token, setToken] = useState<AuthToken | null>(null);
+  const [userId, setUserId] = useState<number | null>(JSON.parse(localStorage.getItem(StorageNames.userId) as string) || null);
+
+  const [rawToken, setRawToken] = useState<AuthToken | null>(null);
+  const confirmMail = JSON.parse(localStorage.getItem(StorageNames.confirmMail) as string);
 
   useEffect(() => {
-    if (token) {
+    if (rawToken) {
+      localStorage.setItem(StorageNames.rawToken, JSON.stringify(rawToken));
+    }
+  }, [rawToken]);
+
+  useEffect(() => {
+    if (token && !confirmMail) {
       onGetUser();
       localStorage.setItem(StorageNames.token, JSON.stringify(token));
     }
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [token]);
 
+
   useEffect(() => {
     const storageTokens = JSON.parse(localStorage.getItem(StorageNames.token) as string);
+    const storageRawTokens = JSON.parse(localStorage.getItem(StorageNames.rawToken) as string);
+    const userId = JSON.parse(localStorage.getItem(StorageNames.userId) as string);
 
+
+
+    if (userId) {
+      setUserId(Number(userId));
+    }
+    if (storageRawTokens) {
+      setRawToken(storageRawTokens)
+    }
     if (storageTokens) {
       setToken(storageTokens);
-    } else {
+    } else if (!confirmMail) {
       onLogout();
     }
   }, []);
@@ -51,7 +73,6 @@ export const JWTProvider = ({ children }: PropsWithChildren) => {
       if (err.code === 'token_not_valid') {
         onRefresh();
       } else {
-        console.log('hello');
         dispatch({
           type: LOGOUT
         });
@@ -77,12 +98,27 @@ export const JWTProvider = ({ children }: PropsWithChildren) => {
   };
 
   const onRegister = async (email: string, password: string): Promise<void> => {
-    const { data: tokens } = await axios.post('/v1/auth/register/', {
+    const { data: tokens } = await axios.post<AuthToken>('/v1/auth/register/', {
       email,
       password
     });
+    const decode = jwt.decode(tokens.access) as { user_id: number };
 
-    setToken(tokens);
+    localStorage.setItem(StorageNames.confirmMail, JSON.stringify(email));
+    localStorage.setItem(StorageNames.userId, JSON.stringify(decode.user_id));
+
+    await axios.post('/v1/users/send-confirmation-to-email/', { user_id: decode.user_id });
+
+    setUserId(decode.user_id);
+    setRawToken(tokens);
+  };
+
+  const onConfirmEmail = async (code: string): Promise<void> => {
+    await axios.post('/v1/users/confirm-email/', { user_id: userId, confirmation_code: code });
+    setToken(rawToken);
+    localStorage.removeItem(StorageNames.rawToken);
+    localStorage.removeItem(StorageNames.confirmMail);
+    localStorage.removeItem(StorageNames.userId);
   };
 
   const onRegisterViaMedia = async (type: keyof typeof SocialMediaType, code: string): Promise<void> => {
@@ -140,10 +176,12 @@ export const JWTProvider = ({ children }: PropsWithChildren) => {
     <JWTContext.Provider
       value={{
         loading,
+        confirmMail,
         user,
         onLogin,
         onRegister,
         onRegisterViaMedia,
+        onConfirmEmail,
         onUpdateUser,
         onUpdateAvatar,
         onLogout
